Permissions: meh
We're almost there, I think.

I'm still running into issues with the READ ONLY bit being set on the
files, which is all PHP seems to check when it performs an is_writable()
check on a directory. Would be worth a look with procmon to see if this
is PHP setting the bit or whether Samba is trying to be clever (see the
map read only option in smb.conf).